Saves checkout URL in PayRequest
Updates the PayRequest entity to store the checkout URL received from the payment provider.
This allows for easier retrieval and use of the checkout URL later on.

Also, modifies the PayRequestService to update the PayRequest with the checkoutUrl from the provider response.

Updates PayRequest DTO to return the checkout url.